+++ mongo_rep_admin.py
@@ -609,7 +609,7 @@ def main():
     file_chk_list = ["-o"]
     file_crt_list = ["-o"]
     func_dict = {"-L": chk_rep_lag, "-M": fetch_members, "-S": chk_rep_stat,
-                 "-P": fetch_priority, "-T": prt_rep_stat}
+                 "-P": fetch_priority, "-T": prt_rep_stat, "-N": node_chk}
     opt_con_req_list = {"-i": ["-m"], "-s": ["-e"]}
     opt_def_dict = {"-j": False, "-i": "sysmon:mongo_rep_lag"}
     opt_multi_list = ["-e", "-s"]