+++ CHANGELOG.md
@@ -12,7 +12,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 ### Changed
 - main: Added -N option to function dictionary.
 - run_program:  Changed configuration entry match configuration file.
-- config/mongo.py.TEMPLATE:  Changed configuration entry.
+- config/mongo.py.TEMPLATE:  Changed configuration entry and added three new entries.
 
 
 ## [3.3.0] - 2020-07-13